+++ swing/framework-model/src/main/java/is/codion/swing/framework/model/SwingEntityApplicationModel.java
@@ -24,6 +24,8 @@ import is.codion.framework.model.DefaultEntityApplicationModel;
 
 import java.util.Collection;
 
+import static java.util.Objects.requireNonNull;
+
 /**
  * A Swing implementation of {@link is.codion.framework.model.EntityApplicationModel}
  */
@@ -36,7 +38,7 @@ public class SwingEntityApplicationModel
 	 * @param entityModels the entity models
 	 */
 	public SwingEntityApplicationModel(EntityConnectionProvider connectionProvider, Collection<SwingEntityModel> entityModels) {
-		this(connectionProvider, entityModels, null);
+		super(connectionProvider, entityModels, null);
 	}
 
 	/**
@@ -46,6 +48,6 @@ public class SwingEntityApplicationModel
 	 * @param version the application version
 	 */
 	public SwingEntityApplicationModel(EntityConnectionProvider connectionProvider, Collection<SwingEntityModel> entityModels, Version version) {
-		super(connectionProvider, entityModels, version);
+		super(connectionProvider, entityModels, requireNonNull(version));
 	}
 }